# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DreamCraft API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comments": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get random comments",
                "operationId": "72c0f4e68696bbef3567eb3bf1cb5ab7",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-04-27T17:45:19.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-04-27T17:45:19.000000Z"
                                                    },
                                                    "text": {
                                                        "type": "string",
                                                        "example": "Placeat vitae deserunt ullam voluptatem sit."
                                                    },
                                                    "is_edit": {
                                                        "type": "string",
                                                        "example": "0"
                                                    },
                                                    "category_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "category_title": {
                                                        "type": "string",
                                                        "example": "aut"
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nickname": {
                                                        "type": "string",
                                                        "example": "mellie79"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "https://i.servimg.com/u/f21/18/21/41/30/na1110.png"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "success": false,
                                            "message": "Something wrong. Try later."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Create comment",
                "operationId": "b51a43a64d8985e7dbdc5dee7e0050d4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "text": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "category_id": 1,
                                    "text": "Text of comment"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create comment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "data",
                                                    "example": "2024-04-27T21:27:04.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "data",
                                                    "example": "2024-04-27T21:27:04.000000Z"
                                                },
                                                "text": {
                                                    "type": "string",
                                                    "example": "Text of comment"
                                                },
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "message": "Unauthenticated."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "success": false,
                                            "data": "Category not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "message": "The title field must be a string.",
                                            "errors": {
                                                "title": "The title field must be a string."
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "success": false,
                                            "message": "Something wrong. Try later."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}